Integer refactor + lazy mode test

+++ src/main/java/com/mcichosz/studentsystem/model/FacultyCourseId.java
@@ -9,14 +9,14 @@ import java.util.Objects;
 public class FacultyCourseId  implements Serializable {
 
     @Column(name = "faculty_id")
-    private Long facultyId;
+    private Integer facultyId;
 
     @Column(name = "course_id")
-    private Long courseId;
+    private Integer courseId;
 
     private FacultyCourseId() {}
 
-    public FacultyCourseId(Long facultyId, Long courseId) {
+    public FacultyCourseId(Integer facultyId, Integer courseId) {
         this.facultyId = facultyId;
         this.courseId = courseId;
     }
